move PageLink animation into its own file
